top: warn of man page impact if 'USE_X_COLHDR' defined
This warning was added after discovering openSUSE uses
the option with (at least) their 'tumbleweed' version.

The man document sections impacted are: 4e, 5d and 5e.
That is where it was asserted sort column highlighting
would be temporarily turned off if search ('L') and/or
filtering ('O') is active. This option makes it false.

[ shame on suse for not also correcting the man page ]

Signed-off-by: Jim Warner <james.warner@comcast.net>