top: bump current upper max to 100 for field additions
As I speculated in a prior commit, this patch will now
demonstrate how simple it is to extend the upper limit
of supportable fields. With these changes we'll now go
from facing the old limit of 86 to a new limit of 100.

No other changes are needed nor is any logic impacted.

Signed-off-by: Jim Warner <james.warner@comcast.net>