build-sys: Repeat man pages into man-po Makefile
I'm not sure why this is needed, but if the dist_man_MANS
variable is not copied from the top-level Makefile down to the
man-po one then make dist from a clean archive will not work.

I'm sure someone with better automake magic than me can make
this more elegant but for now it works if a little ugly.

References:
  https://www.freelists.org/post/procps/important-newlib-patches,8