slabinfo: add comment for cache_size calculation
One of our physical machine shows that the "CACHE SIZE" column
of slabtop output is extremely high, three times of the products
of objs nums and objs size. After some analysis, we found that
the order of slab, which decides "pages per slab", will shrink
when memory pressure is high and normal order allocation failed.
So we think it might help to add these comments to the man help.

Minor fix: add the "memory." back, which is lost after
"aa461df0: docs: Minor manpage fixes"

Signed-off-by: Zhao Mengmeng <zhaomengmeng@kylinos.cn>