top: lessen overhead when field is not being displayed
This program employs fixed size stacks which currently
total 85 result structures. When a field was not being
shown, those unused result structures were initialized
as 'PIDS_extra' items. As a result, that library would
then reset the results to zero with every interaction.

For any of those result structs holding values created
by top this was appropriate. But for most, it was not.

So this patch will reduce that use of PIDS_extra while
vastly increasing use of the less expensive PIDS_noop.

Signed-off-by: Jim Warner <james.warner@comcast.net>